+++ app/middleware/authWithoutBind.ts
@@ -6,11 +6,11 @@ export default defineNuxtRouteMiddleware(() => {
 
   // On server, check if the user isn't authenticated
   // and redirect to /sign-in.
-  if (process.server && !nuxtApp.ssrContext?.event.context.auth?.userId)
+  if (import.meta.server && !nuxtApp.ssrContext?.event.context.auth?.userId)
     return navigateTo('/sign-in')
 
   // On client, check if clerk is loaded and if user isn't authenticated
   // and redirect to /sign-in.
-  if (process.client && clerk.loaded && !clerk.user?.id)
+  if (import.meta.client && clerk.loaded && !clerk.user?.id)
     return navigateTo('/sign-in')
 })