+++ server/api/files/club-records.post.ts
@@ -1,5 +1,10 @@
 import { PrismaClient } from '@prisma/client'
 
+interface Body {
+  clubId: number
+  collection: string
+}
+
 const prisma = new PrismaClient()
 
 export default eventHandler(async (event) => {
@@ -10,12 +15,16 @@ export default eventHandler(async (event) => {
   }
 
   return readBody(event)
-    .then(async (body) => {
+    .then(async (body: Body) => {
       const { clubId, collection } = body
       const records = await prisma.fileUploadRecord.findMany({
         where: {
-          clubId,
-          fileUploadId: collection,
+          clubId: {
+            equals: clubId,
+          },
+          fileUploadId: {
+            equals: collection,
+          },
         },
         include: {
           file: true,